refactor(tests): reorganize test files under each module __tests__
- libs/ai:
  move system and unit tests into ai/__tests__ (run-ai, model-utils)

- libs/file-io:
  move system, functional, and unit tests into file-io/__tests__
  update import paths in backup and walk-files tests

- libs/io:
  move unit tests into io/__tests__ (hash, logger, parse-args)
  update import path in logger test

- libs/parallel:
  move unit test into parallel/__tests__ (concurrency)

- libs/text:
  move all unit tests into text/__tests__ (date, frontmatter, json, line, markdown, slug, string)

- overall:
  relocate tests from shared __tests__ to module-local __tests__ structure
