+++ skills/_scripts/libs/parallel/__tests__/unit/concurrency.unit.spec.ts
@@ -1,0 +1,220 @@
+// src: skills/_scripts/libs/__tests__/unit/concurrency.unit.spec.ts
+// @(#): withConcurrency / createTasks / runConcurrent / runChunked のユニットテスト
+//
+// Copyright (c) 2026- atsushifx <https://github.com/atsushifx>
+//
+// This software is released under the MIT License.
+// https://opensource.org/licenses/MIT
+
+// -- BDD modules --
+import { assertEquals, assertRejects } from '@std/assert';
+import { describe, it } from '@std/testing/bdd';
+
+// -- test target --
+import { createTasks, runChunked, runConcurrent, withConcurrency } from '../../concurrency.ts';
+
+// ─────────────────────────────────────────────
+// withConcurrency
+// ─────────────────────────────────────────────
+
+describe('withConcurrency', () => {
+  describe('Given: tasks 3個, limit 2', () => {
+    describe('When: withConcurrency を実行する', () => {
+      describe('Then: T-LIB-C-01 - 結果が入力順で返る', () => {
+        it('T-LIB-C-01-01: 結果配列が入力順序と一致する', async () => {
+          const _tasks = [1, 2, 3].map((n) => () => Promise.resolve(n * 10));
+          const _results = await withConcurrency(_tasks, 2);
+          assertEquals(_results, [10, 20, 30]);
+        });
+      });
+    });
+  });
+
+  describe('Given: tasks 2個, limit 5 (limit > tasks.length)', () => {
+    describe('When: withConcurrency を実行する', () => {
+      describe('Then: T-LIB-C-02 - 正常に結果を返す', () => {
+        it('T-LIB-C-02-01: worker 数が tasks.length を超えても正常に返る', async () => {
+          const _tasks = [1, 2].map((n) => () => Promise.resolve(n));
+          const _results = await withConcurrency(_tasks, 5);
+          assertEquals(_results, [1, 2]);
+        });
+      });
+    });
+  });
+
+  describe('Given: tasks が空配列', () => {
+    describe('When: withConcurrency を実行する', () => {
+      describe('Then: T-LIB-C-03 - 空配列を返す', () => {
+        it('T-LIB-C-03-01: 空の結果配列が返る', async () => {
+          const _results = await withConcurrency([], 4);
+          assertEquals(_results, []);
+        });
+      });
+    });
+  });
+
+  describe('Given: tasks のうち1個が reject する', () => {
+    describe('When: withConcurrency を実行する', () => {
+      describe('Then: T-LIB-C-04 - reject が伝播する', () => {
+        it('T-LIB-C-04-01: reject した Error が伝播する', async () => {
+          const _tasks = [
+            () => Promise.resolve(1),
+            () => Promise.reject(new Error('task failed')),
+            () => Promise.resolve(3),
+          ];
+          await assertRejects(
+            () => withConcurrency(_tasks, 2),
+            Error,
+            'task failed',
+          );
+        });
+      });
+    });
+  });
+});
+
+// ─────────────────────────────────────────────
+// createTasks
+// ─────────────────────────────────────────────
+
+describe('createTasks', () => {
+  describe('Given: items 3個, fn 定義済み', () => {
+    describe('When: createTasks を実行する', () => {
+      describe('Then: T-LIB-C-05 - Task[] が返る', () => {
+        it('T-LIB-C-05-01: 返値の長さが items.length と一致する', () => {
+          const _tasks = createTasks([1, 2, 3], (n) => Promise.resolve(n));
+          assertEquals(_tasks.length, 3);
+        });
+
+        it('T-LIB-C-05-02: 各要素が関数である', () => {
+          const _tasks = createTasks([1, 2, 3], (n) => Promise.resolve(n));
+          for (const task of _tasks) {
+            assertEquals(typeof task, 'function');
+          }
+        });
+      });
+    });
+  });
+
+  describe('Given: createTasks の Task[] を withConcurrency で実行', () => {
+    describe('When: withConcurrency に渡す', () => {
+      describe('Then: T-LIB-C-06 - 全 item に fn 適用した結果が返る', () => {
+        it('T-LIB-C-06-01: 各 item に fn を適用した結果配列が返る', async () => {
+          const _tasks = createTasks(['a', 'b', 'c'], (s) => Promise.resolve(s.toUpperCase()));
+          const _results = await withConcurrency(_tasks, 2);
+          assertEquals(_results, ['A', 'B', 'C']);
+        });
+      });
+    });
+  });
+
+  describe('Given: items が空配列', () => {
+    describe('When: createTasks を実行する', () => {
+      describe('Then: T-LIB-C-07 - 空の Task[] を返す', () => {
+        it('T-LIB-C-07-01: 空の配列が返る', () => {
+          const _tasks = createTasks([], (n: number) => Promise.resolve(n));
+          assertEquals(_tasks, []);
+        });
+      });
+    });
+  });
+});
+
+// ─────────────────────────────────────────────
+// runConcurrent
+// ─────────────────────────────────────────────
+
+describe('runConcurrent', () => {
+  describe('Given: items 3個, fn 定義済み, limit=2', () => {
+    describe('When: runConcurrent を実行する', () => {
+      describe('Then: T-LIB-C-08 - 入力順の結果配列が返る', () => {
+        it('T-LIB-C-08-01: 結果配列が入力順序と一致する', async () => {
+          const _results = await runConcurrent([1, 2, 3], (n) => Promise.resolve(n * 10), 2);
+          assertEquals(_results, [10, 20, 30]);
+        });
+      });
+    });
+  });
+
+  describe('Given: items が空配列', () => {
+    describe('When: runConcurrent を実行する', () => {
+      describe('Then: T-LIB-C-09 - 空配列を返す', () => {
+        it('T-LIB-C-09-01: 空の結果配列が返る', async () => {
+          const _results = await runConcurrent([], (n: number) => Promise.resolve(n), 2);
+          assertEquals(_results, []);
+        });
+      });
+    });
+  });
+
+  describe('Given: fn が reject する', () => {
+    describe('When: runConcurrent を実行する', () => {
+      describe('Then: T-LIB-C-10 - reject が伝播する', () => {
+        it('T-LIB-C-10-01: reject した Error が伝播する', async () => {
+          await assertRejects(
+            () =>
+              runConcurrent(
+                [1, 2, 3],
+                (n) => n === 2 ? Promise.reject(new Error('item failed')) : Promise.resolve(n),
+                2,
+              ),
+            Error,
+            'item failed',
+          );
+        });
+      });
+    });
+  });
+});
+
+// ─────────────────────────────────────────────
+// runChunked
+// ─────────────────────────────────────────────
+
+describe('runChunked', () => {
+  describe('Given: items 5個, chunkSize=2, limit=2', () => {
+    describe('When: runChunked を実行する', () => {
+      describe('Then: T-LIB-C-11 - chunk 単位で fn が呼ばれ結果が返る', () => {
+        it('T-LIB-C-11-01: fn が chunk 配列を受け取り、結果配列が返る', async () => {
+          const _received: number[][] = [];
+          const _fn = (chunk: number[]): Promise<number> => {
+            _received.push(chunk);
+            return Promise.resolve(chunk.reduce((a, b) => a + b, 0));
+          };
+          const _results = await runChunked([1, 2, 3, 4, 5], 2, _fn, 2);
+          assertEquals(_received, [[1, 2], [3, 4], [5]]);
+          assertEquals(_results, [3, 7, 5]);
+        });
+      });
+    });
+  });
+
+  describe('Given: items が空配列', () => {
+    describe('When: runChunked を実行する', () => {
+      describe('Then: T-LIB-C-12 - 空配列を返す', () => {
+        it('T-LIB-C-12-01: 空の結果配列が返る', async () => {
+          const _results = await runChunked([], 2, (chunk: number[]) => Promise.resolve(chunk.length), 2);
+          assertEquals(_results, []);
+        });
+      });
+    });
+  });
+
+  describe('Given: chunkSize が items.length より大きい', () => {
+    describe('When: runChunked を実行する', () => {
+      describe('Then: T-LIB-C-13 - 1 chunk で fn が呼ばれる', () => {
+        it('T-LIB-C-13-01: fn が全 items を含む 1 chunk を受け取る', async () => {
+          const _received: number[][] = [];
+          const _fn = (chunk: number[]): Promise<number> => {
+            _received.push(chunk);
+            return Promise.resolve(chunk.length);
+          };
+          const _results = await runChunked([1, 2, 3], 10, _fn, 2);
+          assertEquals(_received.length, 1);
+          assertEquals(_received[0], [1, 2, 3]);
+          assertEquals(_results, [3]);
+        });
+      });
+    });
+  });
+});
